raspberry/requirements.txt
RPi.GPIO
flask
python-crontab
